+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 1.1.0 (In development)
 ### New Features
+* Added `GridView` component
 * Items can now be removed via index with `ListView.removeAt`
-* The scrollbar can now optionally be hidden when there are no items to display by passing `hideWhenEmpty: true` to `setScrollbarEnabled`
+* The scrollbar can now be hidden when there are no items to display by passing `hideWhenEmpty: true` to `setScrollbarEnabled`
 
 ### Bug Fixes
 * Restored `Listview.remove` after its removal in the 1.0.0 refactor